Extract separate modules from `TestChain`: step 3/3
