+++ eras/shelley/test-suite/src/Test/Cardano/Ledger/Shelley/Rules/Pool.hs
@@ -1,34 +1,139 @@
+{-# LANGUAGE AllowAmbiguousTypes #-}
 {-# LANGUAGE DataKinds #-}
+{-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
+{-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
 
 module Test.Cardano.Ledger.Shelley.Rules.Pool (
-  poolRegistration,
-  poolStateIsInternallyConsistent,
-  poolRetirement,
+  poolProps,
 )
 where
 
-import Cardano.Ledger.Shelley.Delegation.Certificates (PoolCert (RegPool, RetirePool))
+import Cardano.Ledger.Block (
+  bheader,
+ )
+import Cardano.Ledger.Core
+import Cardano.Ledger.Shelley.Core
 import Cardano.Ledger.Shelley.LedgerState (
+  EpochState (..),
+  NewEpochState (..),
   PState (..),
   psFutureStakePoolParams,
   psStakePoolParams,
  )
 import Cardano.Ledger.Shelley.Rules (ShelleyPOOL)
-import Cardano.Ledger.Shelley.TxBody (DCert (DCertPool), PoolParams (..))
+import Cardano.Ledger.Shelley.TxBody
 import Cardano.Ledger.Slot (EpochNo (..))
+import Cardano.Protocol.TPraos.BHeader (
+  bhbody,
+  bheaderSlotNo,
+ )
 import Control.SetAlgebra (dom, eval, (∈), (∉))
 import Control.State.Transition.Trace (
   SourceSignalTarget (..),
+  TraceOrder (OldestFirst),
+  sourceSignalTargets,
+  traceStates,
  )
-import qualified Data.Map.Strict as Map (keysSet, lookup)
-import qualified Data.Set as Set (isSubsetOf)
-import Test.QuickCheck (Property, conjoin, counterexample, property, (===))
+import qualified Control.State.Transition.Trace.Generator.QuickCheck as QC
+import qualified Data.Map.Strict as Map
+import qualified Data.Set as Set
+import Lens.Micro.Extras (view)
+import Test.Cardano.Ledger.Shelley.Generator.Core (GenEnv)
+import Test.Cardano.Ledger.Shelley.Generator.EraGen (EraGen (..))
+import Test.Cardano.Ledger.Shelley.Generator.ShelleyEraGen ()
+import Test.Cardano.Ledger.Shelley.Rules.Chain (CHAIN, ChainState (..))
+import Test.QuickCheck (Property, conjoin, counterexample, (===))
 
-poolRegistration :: SourceSignalTarget (ShelleyPOOL era) -> Property
-poolRegistration
+import Test.Cardano.Ledger.Shelley.Rules.TestChain (
+  forAllChainTrace,
+  poolTraceFromBlock,
+  traceLen,
+ )
+import Test.Cardano.Ledger.Shelley.Utils (
+  ChainProperty,
+  epochFromSlotNo,
+ )
+import Test.QuickCheck (
+  Testable (..),
+ )
+
+-- | Various properties of the POOL STS Rule, tested on longer traces
+-- (double the default length)
+poolProps ::
+  forall era.
+  ( EraGen era
+  , EraGovernance era
+  , ChainProperty era
+  , QC.HasTrace (CHAIN era) (GenEnv era)
+  , ProtVerAtMost era 8
+  ) =>
+  Property
+poolProps =
+  forAllChainTrace @era traceLen $ \tr -> do
+    let ssts = sourceSignalTargets tr
+    conjoin . concat $
+      [ map poolRetirement ssts
+      , map poolRegistration ssts
+      , map poolStateIsInternallyConsistent ssts
+      ]
+
+-- | Check that a `RetirePool` certificate properly marks a stake pool for
+-- retirement.
+poolRetirement ::
+  ( ChainProperty era
+  , EraSegWits era
+  , ShelleyEraTxBody era
+  , ProtVerAtMost era 8
+  ) =>
+  SourceSignalTarget (CHAIN era) ->
+  Property
+poolRetirement SourceSignalTarget {source = chainSt, signal = block} =
+  conjoin $
+    map (poolRetirementProp currentEpoch maxEpoch) (sourceSignalTargets poolTr)
+  where
+    (chainSt', poolTr) = poolTraceFromBlock chainSt block
+    bhb = bhbody $ bheader block
+    currentEpoch = (epochFromSlotNo . bheaderSlotNo) bhb
+    maxEpoch = (view ppEMaxL . esPp . nesEs . chainNes) chainSt'
+
+-- | Check that a newly registered pool key is registered and not
+-- in the retiring map.
+poolRegistration ::
+  ( ChainProperty era
+  , EraSegWits era
+  , ShelleyEraTxBody era
+  , ProtVerAtMost era 8
+  ) =>
+  SourceSignalTarget (CHAIN era) ->
+  Property
+poolRegistration (SourceSignalTarget {source = chainSt, signal = block}) =
+  conjoin $
+    map poolRegistrationProp (sourceSignalTargets poolTr)
+  where
+    (_, poolTr) = poolTraceFromBlock chainSt block
+
+-- | Assert that PState maps are in sync with each other after each `Signal
+-- POOL` transition.
+poolStateIsInternallyConsistent ::
+  ( ChainProperty era
+  , EraSegWits era
+  , ShelleyEraTxBody era
+  , ProtVerAtMost era 8
+  ) =>
+  SourceSignalTarget (CHAIN era) ->
+  Property
+poolStateIsInternallyConsistent (SourceSignalTarget {source = chainSt, signal = block}) =
+  conjoin $
+    map poolStateIsInternallyConsistentProp (traceStates OldestFirst poolTr)
+  where
+    (_, poolTr) = poolTraceFromBlock chainSt block
+
+poolRegistrationProp :: SourceSignalTarget (ShelleyPOOL era) -> Property
+poolRegistrationProp
   SourceSignalTarget
     { signal = (DCertPool (RegPool poolParams))
     , source = sourceSt
@@ -62,10 +167,10 @@ poolRegistration
                   "PoolParams are removed in 'retiring'"
                   (eval (hk ∉ dom (psRetiring targetSt)) :: Bool)
               ]
-poolRegistration _ = property ()
+poolRegistrationProp _ = property ()
 
-poolRetirement :: EpochNo -> EpochNo -> SourceSignalTarget (ShelleyPOOL era) -> Property
-poolRetirement
+poolRetirementProp :: EpochNo -> EpochNo -> SourceSignalTarget (ShelleyPOOL era) -> Property
+poolRetirementProp
   currentEpoch@(EpochNo ce)
   (EpochNo maxEpoch)
   SourceSignalTarget {source = sourceSt, target = targetSt, signal = (DCertPool (RetirePool hk e))} =
@@ -83,10 +188,10 @@ poolRetirement
           "hk must be in target's retiring"
           (eval (hk ∈ dom (psRetiring targetSt)) :: Bool)
       ]
-poolRetirement _ _ _ = property ()
+poolRetirementProp _ _ _ = property ()
 
-poolStateIsInternallyConsistent :: PState c -> Property
-poolStateIsInternallyConsistent PState {psStakePoolParams = pParams_, psRetiring = retiring_} = do
+poolStateIsInternallyConsistentProp :: PState c -> Property
+poolStateIsInternallyConsistentProp PState {psStakePoolParams = pParams_, psRetiring = retiring_} = do
   let poolKeys = Map.keysSet pParams_
       pParamKeys = Map.keysSet pParams_
       retiringKeys = Map.keysSet retiring_
